+++ app/Http/Middleware/EnsureModelNotStale.php
@@ -3,6 +3,7 @@
 namespace App\Http\Middleware;
 
 use App\Enums\CustomStatusCodes;
+use Carbon\Carbon;
 use Closure;
 use Illuminate\Http\Request;
 
@@ -13,13 +14,14 @@ class EnsureModelNotStale
      * request parameters is older than the updated_at of the model in the database, a error response with the
      * status code 466 will be returned containing the actual model.
      *
-     * @param  Request $request       Request to get the model and parameters from
-     * @param  Closure $next          Next to call if everything is ok
-     * @param  String  $parameterName Name of the parameter to get the model from
-     * @param  String  $resourceClass Resource class to cast the new model with
+     * @param  Request $request                Request to get the model and parameters from
+     * @param  Closure $next                   Next to call if everything is ok
+     * @param  String  $parameterName          Name of the parameter to get the model from
+     * @param  String  $resourceClass          Resource class to cast the new model with
+     * @param  mixed   ...$relationshipsToLoad Additional relationships to load for the updated model in the response
      * @return mixed
      */
-    public function handle($request, Closure $next, $parameterName, $resourceClass)
+    public function handle($request, Closure $next, $parameterName, $resourceClass, ...$relationshipsToLoad)
     {
         $request->validate([
             'updated_at' => 'required|date'
@@ -27,10 +29,12 @@ class EnsureModelNotStale
 
         $model = $request->route($parameterName);
 
-        if ($request->updated_at < $model->updated_at) {
+        if ($model->updated_at->gt(Carbon::parse($request->updated_at))) {
+            $model->load($relationshipsToLoad);
+
             return response()->json([
                 'error'     => CustomStatusCodes::STALE_MODEL,
-                'message'   => trans('app.errors.stale_model', ['model' => trans('app.model.' . $model->getTable())]),
+                'message'   => __('app.errors.stale_model', ['model' => __('app.model.' . $model->getTable())]),
                 'new_model' => new $resourceClass($model)
             ], CustomStatusCodes::STALE_MODEL);
         }