+++ resources/js/components/Permissions/Cannot.vue
@@ -21,7 +21,7 @@ import { Fragment } from 'vue-fragment';
  *     Hello World!
  *   </cannot>
  *   <!-- or -->
- *   <cannot method="viewAll" :policy="{ modelName: 'User', id: 1 ... }">
+ *   <cannot method="viewAll" :policy="{ model_name: 'User', id: 1 ... }">
  *     Hello World!
  *   </cannot>
  *   ```