+++ resources/js/components/Room/FileComponent.vue
@@ -14,8 +14,7 @@
         >
         </b-form-file>
 
-        <b-form-invalid-feedback>
-          {{ fieldError('file') }}
+        <b-form-invalid-feedback v-html="fieldError('file')">
         </b-form-invalid-feedback>
 
         <b-form-text>{{ $t('rooms.files.formats',{formats: files.file_mimes}) }}<br>{{ $t('rooms.files.size',{size: files.file_size}) }}</b-form-text>
@@ -120,8 +119,10 @@
 </template>
 <script>
 import Base from '../../api/base';
+import FieldErrors from '../../mixins/FieldErrors';
 
 export default {
+  mixins: [FieldErrors],
   props: {
     room: Object
   },
@@ -138,15 +139,6 @@ export default {
     };
   },
   methods: {
-
-    fieldState (field) {
-      return this.errors[field] === undefined ? null : false;
-    },
-    fieldError (field) {
-      if (this.fieldState(field) !== false) { return ''; }
-      return this.errors[field].join('<br>');
-    },
-
     /**
      * Request file download url
      * @param file file object