+++ resources/js/lang/en/settings.js
@@ -4,7 +4,31 @@ export default {
     'Please select one of the menu items on the left to adjust the settings.',
 
   roles: {
-    title: 'Roles'
+    title: 'Roles',
+
+    new: 'Create new role',
+    view: 'Detailed information for {name}',
+    edit: 'Edit role {name}',
+
+    id: 'ID',
+    name: 'Name',
+    permissions: 'Permissions',
+    roomLimit: {
+      label: 'Room limit',
+      default: 'System default ({value})',
+      unlimited: 'Unlimited',
+      custom: 'Custom amount'
+    },
+    default: 'Default',
+    actions: 'Actions',
+    nodata: 'No roles found!',
+
+    noOptions: 'No permissions found!',
+
+    delete: {
+      confirm: 'Are you really want to delete the role {name}?',
+      title: 'Delete role?'
+    }
   },
 
   users: {