+++ resources/lang/en/app.php
@@ -9,8 +9,10 @@ return [
         'file_not_found' => 'The file could not be found.',
         'no_server_available' => 'Currently there are no servers available.',
         'room_limit_exceeded' => 'Creating room failed! You have reached the max. amount of rooms.',
+        'role_delete_linked_users' => 'The role is linked to users and therefore it can\'t be deleted!',
+        'role_update_permission_lost' => 'The changes of the role would lead to a lost of permissions to view and update roles for you!',
     ],
     'model' => [
         'roles' => 'role'
-    ],
+    ]
 ];