+++ tests/Frontend/Router.spec.js
@@ -1,6 +1,21 @@
-import { beforeEachRoute } from '../../resources/js/router';
+import { beforeEachRoute, routes } from '../../resources/js/router';
+import PermissionService from '../../resources/js/services/PermissionService';
+import moxios from 'moxios';
+import sinon from 'sinon';
+import Base from '../../resources/js/api/base';
+
+const accessPermittedRolesView = routes.filter(route => route.path === '/settings')[0]
+  .children.filter(route => route.name === 'settings.roles.view')[0].meta.accessPermitted;
 
 describe('Router', function () {
+  beforeEach(function () {
+    moxios.install();
+  });
+
+  afterEach(function () {
+    moxios.uninstall();
+  });
+
   describe('beforeEachRoute', function () {
     it('beforeEachRoute calls next if there is no permission checks or required authentication', function (done) {
       const router = {};
@@ -11,7 +26,8 @@ describe('Router', function () {
         },
         state: {
           initialized: true
-        }
+        },
+        commit: () => {}
       };
 
       const to = {
@@ -40,7 +56,8 @@ describe('Router', function () {
         },
         state: {
           initialized: true
-        }
+        },
+        commit: () => {}
       };
 
       const to = {
@@ -79,7 +96,8 @@ describe('Router', function () {
         },
         state: {
           initialized: true
-        }
+        },
+        commit: () => {}
       };
 
       const to = {
@@ -103,6 +121,130 @@ describe('Router', function () {
           expect(errors[1]).toBe('app.flash.unauthorized');
           done();
         });
+      });
+    });
+  });
+
+  describe('accessPermitted', function () {
+    it('for role detail view returns true if user has the necessary permissions', function (done) {
+      const oldUser = PermissionService.currentUser;
+
+      accessPermittedRolesView({ id: 1 }, { view: '1' }).then(result => {
+        expect(result).toBe(false);
+
+        PermissionService.setCurrentUser({ permissions: ['roles.view'] });
+        return accessPermittedRolesView({ id: 1 }, { view: '1' });
+      }).then(result => {
+        expect(result).toBe(false);
+
+        PermissionService.setCurrentUser({ permissions: ['roles.view', 'settings.manage'] });
+        return accessPermittedRolesView({ id: 1 }, { view: '1' });
+      }).then(result => {
+        expect(result).toBe(true);
+
+        PermissionService.setCurrentUser(oldUser);
+        done();
+      });
+    });
+
+    it('for role new view returns true if user has the necessary permissions', function (done) {
+      const oldUser = PermissionService.currentUser;
+
+      accessPermittedRolesView({ id: 'new' }, {}).then(result => {
+        expect(result).toBe(false);
+
+        PermissionService.setCurrentUser({ permissions: ['roles.create'] });
+        return accessPermittedRolesView({ id: 'new' }, {});
+      }).then(result => {
+        expect(result).toBe(false);
+
+        PermissionService.setCurrentUser({ permissions: ['roles.create', 'settings.manage'] });
+        return accessPermittedRolesView({ id: 'new' }, {});
+      }).then(result => {
+        expect(result).toBe(true);
+
+        PermissionService.setCurrentUser(oldUser);
+        done();
+      });
+    });
+
+    it('for role update view returns true if user has necessary permissions and role is not a default role', function (done) {
+      const oldUser = PermissionService.currentUser;
+      const respond = function () {
+        const request = moxios.requests.mostRecent();
+        request.respondWith({
+          status: 200,
+          response: {
+            data: {
+              id: 1,
+              default: true,
+              model_name: 'Role'
+            }
+          }
+        });
+      };
+
+      moxios.wait(respond);
+
+      accessPermittedRolesView({ id: 1 }, {}).then(result => {
+        expect(result).toBe(false);
+
+        PermissionService.setCurrentUser({ permissions: ['roles.update'] });
+        moxios.wait(respond);
+
+        return accessPermittedRolesView({ id: 1 }, {});
+      }).then(result => {
+        expect(result).toBe(false);
+
+        PermissionService.setCurrentUser({ permissions: ['roles.update', 'settings.manage'] });
+        moxios.wait(respond);
+
+        return accessPermittedRolesView({ id: 1 }, {});
+      }).then(result => {
+        expect(result).toBe(false);
+
+        moxios.wait(function () {
+          const request = moxios.requests.mostRecent();
+          return request.respondWith({
+            status: 200,
+            response: {
+              data: {
+                id: 1,
+                default: false,
+                model_name: 'Role'
+              }
+            }
+          });
+        });
+
+        return accessPermittedRolesView({ id: 1 }, {});
+      }).then(result => {
+        expect(result).toBe(true);
+
+        PermissionService.setCurrentUser(oldUser);
+        done();
+      });
+    });
+
+    it('for role update view calls error handler returns false on error in request', function (done) {
+      const spy = sinon.spy();
+      sinon.stub(Base, 'error').callsFake(spy);
+
+      moxios.wait(function () {
+        const request = moxios.requests.mostRecent();
+        request.respondWith({
+          status: 500,
+          response: {
+            message: 'Test'
+          }
+        });
+      });
+
+      accessPermittedRolesView({ id: 1 }, {}).then(result => {
+        expect(result).toBe(false);
+        sinon.assert.calledOnce(Base.error);
+        Base.error.restore();
+        done();
       });
     });
   });